requirements.txt
fastapi>=0.115.0
kafka-python>=2.2.0
uvicorn>=0.34.0
pydantic>=2.11.0
jsonschema>=4.23.0
